UPDATED: refactor TapTools pricing integration by renaming client and module, and introducing TapToolsClient for improved token pricing and LP pool data handling
