+++ src/modules/taptools/interfaces/taptools.interface.ts
@@ -1,0 +1,16 @@
+/**
+ * TapTools LP Pool data structure
+ * Represents a liquidity pool from TapTools API
+ */
+export interface TapToolsTokenPoolDto {
+  exchange: string;
+  lpTokenUnit: string;
+  onchainID: string;
+  tokenA: string;
+  tokenALocked: number;
+  tokenATicker: string;
+  /** empty for ADA */
+  tokenB: string | null;
+  tokenBLocked: number;
+  tokenBTicker: string;
+}
