+++ src/modules/vaults/phase-management/governance/governance.module.ts
@@ -35,7 +35,7 @@ import { DexHunterModule } from '@/modules/dexhunter/dexhunter.module';
 import { DistributionCalculationModule } from '@/modules/distribution/distribution-calculation.module';
 import { RedisModule } from '@/modules/redis/redis.module';
 import { RewardsModule } from '@/modules/rewards/rewards.module';
-import { TaptoolsModule } from '@/modules/taptools/taptools.module';
+import { TapToolsPricingModule } from '@/modules/taptools/taptools-pricing.module';
 import { AssetsModule } from '@/modules/vaults/assets/assets.module';
 import { VyfiModule } from '@/modules/vyfi/vyfi.module';
 import { WayUpPricingModule } from '@/modules/wayup/wayup-pricing.module';
@@ -67,7 +67,7 @@ import { WayUpModule } from '@/modules/wayup/wayup.module';
     DistributionCalculationModule,
     RewardsModule,
     WayUpPricingModule,
-    TaptoolsModule,
+    TapToolsPricingModule,
   ],
   controllers: [GovernanceController, TerminationController],
   providers: [
